Refactor handler to support timestamped directories
- Updated migration file naming convention to use timestamps (e.g., `YYYYMMDD_HHMMSS_name`).
- Added support for directory-based migrations with `up` and `down` files.
- Removed legacy `Type` enum and related logic.
- Updated parsing and validation logic for migration files.
- Adjusted tests to reflect new migration structure.

+++ refinery_core/src/drivers/mysql_async.rs
@@ -16,7 +16,7 @@ async fn query_applied_migrations<'a>(
     let applied = result
         .into_iter()
         .map(|row| {
-            let (version, name, applied_on, checksum): (i32, String, String, String) =
+            let (version, name, applied_on, checksum): (i64, String, String, String) =
                 mysql_async::from_row(row);
 
             // Safe to call unwrap, as we stored it in RFC3339 format on the database
